requirements.txt
fastapi
uvicorn
pymongo
cloudinary
python-dotenv
python-multipart